+++ src/commands/runAnonymousApex.ts
@@ -108,25 +108,35 @@ System.debug('sfdx-hardis rocks !!!');
         message: string,
         logFile: string,
       ) => {
-        const openLogAction = "Open Raw Log";
-        const openAnalysisLog = "Open Log Analysis";
+        const openLogAction = "Open Log";
+        const openLogActionDebugOnly = "Open Log (debug only)"
+        const openAnalysisLog = "Open Log (Analysis)";
         let selection: string | undefined;
         if (type === "info") {
           selection = await vscode.window.showInformationMessage(
             message,
             openLogAction,
+            openLogActionDebugOnly,
             openAnalysisLog
           );
         } else {
           selection = await vscode.window.showErrorMessage(
             message,
             openLogAction,
+            openLogActionDebugOnly,
             openAnalysisLog
           );
         }
         if (selection === openLogAction) {
           const document = await vscode.workspace.openTextDocument(logFile);
           await vscode.window.showTextDocument(document);
+        }
+        else if (selection === openLogActionDebugOnly) {
+          const logFileVsCodeUri = vscode.Uri.file(logFile);
+          vscode.commands.executeCommand(
+            "vscode-sfdx-hardis.displayLogDebugOnly",
+            logFileVsCodeUri,
+          );
         }
         else if (selection === openAnalysisLog) {
           // Check if the Lana extension command is available
@@ -187,5 +197,24 @@ System.debug('sfdx-hardis rocks !!!');
       }
     },
   );
+  commands.disposables.push(disposable);
+}
+
+
+export async function registerDisplayLogDebugOnly(commands: Commands) {
+  const disposable = vscode.commands.registerCommand(
+    "vscode-sfdx-hardis.displayLogDebugOnly",
+    async (logFileVsCodeUri: vscode.Uri) => {
+      const logFile = logFileVsCodeUri.fsPath;
+      const debugLogFile = logFile.replace(".log",".debug.log");
+      const logContent = await fs.readFile(logFile, "utf8");
+      const debugLines = logContent
+        .split("\n")
+        .filter((line) => line.includes("|USER_DEBUG|"))
+        .join("\n");
+      await fs.writeFile(debugLogFile, debugLines, "utf8");
+      const document = await vscode.workspace.openTextDocument(debugLogFile);
+      await vscode.window.showTextDocument(document);
+    });
   commands.disposables.push(disposable);
 }